improved typescript export, #357
